Moved BitmapWorkerTask.java and BitmapCache.java
