+++ app/src/main/java/jp/ac/it_college/std/ikemen/reachable/company/coupon/adapter/CouponAdapter.java
@@ -14,8 +14,8 @@ import android.widget.ImageView;
 import java.util.ArrayList;
 import java.util.List;
 
-import jp.ac.it_college.std.ikemen.reachable.company.BitmapCache;
-import jp.ac.it_college.std.ikemen.reachable.company.BitmapWorkerTask;
+import jp.ac.it_college.std.ikemen.reachable.company.coupon.bitmap.BitmapCache;
+import jp.ac.it_college.std.ikemen.reachable.company.coupon.bitmap.BitmapWorkerTask;
 import jp.ac.it_college.std.ikemen.reachable.company.R;
 import jp.ac.it_college.std.ikemen.reachable.company.coupon.CouponFilter;
 import jp.ac.it_college.std.ikemen.reachable.company.info.CouponInfo;
